pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scrub"
version = "0.1.0"
requires-python = ">=3.10,<4"
dependencies = [
    "aiofiles>=23.0",
    "pillow>=12.2.0",
    "PyMuPDF>=1.23",
]

[project.scripts]
scrub = "scrub.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["scrub"]
